Files that conform to the new convention are renamed.
